# internal/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hypervisor orchestration API for managing staged releases, deployments, and hyperuser access tooling.",
        "title": "OpenHack Hypervisor API",
        "contact": {},
        "version": "25.10.17.4"
    },
    "basePath": "/hypervisor",
    "paths": {
        "/hypervisor/deployments": {
            "post": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Deployments"
                ],
                "summary": "Promote stage to deployment",
                "parameters": [
                    {
                        "description": "Promotion payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.promoteStageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Deployment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._DeploymentInvalidRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageNotFound"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageMissingEnv"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/hypervisor/stages": {
            "get": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "List stages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listStagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "description": "Bootstraps a stage by cloning the release repo and seeding the template environment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "Prepare stage",
                "parameters": [
                    {
                        "description": "Stage details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createStageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.StageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageInvalidRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageReleaseNotFound"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageAlreadyExists"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
        "/hypervisor/stages/{stageId}": {
            "get": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "Get stage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stage identifier",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
        "/hypervisor/stages/{stageId}/env": {
            "get": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "Get stage env",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stage identifier",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StageEnvResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "Update stage env",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stage identifier",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Env payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateStageEnvRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageInvalidRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
        "/hypervisor/stages/{stageId}/tests": {
            "post": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Stages"
                ],
                "summary": "Start stage test",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stage identifier",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.StageTestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageInvalidRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._StageNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.StageEnvResponse": {
            "type": "object",
            "properties": {
                "envText": {
                    "type": "string"
                }
            }
        },
        "api.StageResponse": {
            "type": "object",
            "properties": {
                "envText": {
                    "type": "string"
                },
                "stage": {
                    "$ref": "#/definitions/models.Stage"
                }
            }
        },
...
        "api.UpdateStageEnvRequest": {
            "type": "object",
            "properties": {
                "envText": {
                    "type": "string"
                }
            }
        },
        "api.createStageRequest": {
            "type": "object",
            "properties": {
                "envTag": {
                    "type": "string"
                },
                "releaseId": {
                    "type": "string"
                }
            }
        },
...
        "api.listStagesResponse": {
            "type": "object",
            "properties": {
                "stages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Stage"
                    }
                }
            }
        },
        "api.promoteStageRequest": {
            "type": "object",
            "properties": {
                "stageId": {
                    "type": "string"
                }
            }
        },
...
        "errmsg._DeploymentInvalidRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid deployment request payload"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "errmsg._InternalServerError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "internal server error: \u003cdetails\u003e"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "errmsg._StageAlreadyExists": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "stage already exists"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 409
                }
            }
        },
        "errmsg._StageInvalidRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid stage request payload"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "errmsg._StageMissingEnv": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "stage is missing an environment update"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 409
                }
            }
        },
        "errmsg._StageNotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "stage not found"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "errmsg._StageReleaseNotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "release not found"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
...
        "models.Deployment": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "envTag": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "promotedAt": {
                    "type": "string"
                },
                "stageId": {
                    "type": "string"
                },
                "status": {
                    "description": "staged|ready|stopped|deleted",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "models.Stage": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "envTag": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastTestResultId": {
                    "type": "string"
                },
                "releaseId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StageStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.StageStatus": {
            "type": "string",
            "enum": [
                "pre",
                "active",
                "promoted"
            ],
            "x-enum-varnames": [
                "StageStatusPre",
                "StageStatusActive",
                "StageStatusPromoted"
            ]
        },
        "models.StageTestResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logPath": {
                    "type": "string"
                },
                "stageId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StageTestStatus"
                },
                "wsToken": {
                    "type": "string"
                }
            }
        },
        "models.StageTestStatus": {
            "type": "string",
            "enum": [
                "running",
                "passed",
                "failed",
                "canceled",
                "error"
            ],
            "x-enum-varnames": [
                "StageTestStatusRunning",
                "StageTestStatusPassed",
                "StageTestStatusFailed",
                "StageTestStatusCanceled",
                "StageTestStatusError"
            ]
        }
    },
    "securityDefinitions": {
        "HyperUserAuth": {
            "description": "Provide the hyperuser bearer token as `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}